requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
pydantic-settings==2.3.4

# LlamaIndex core (latest)
llama-index==0.11.11
llama-index-readers-file==0.2.2
llama-index-llms-openai==0.2.9
llama-index-embeddings-openai==0.2.5
llama-index-embeddings-huggingface==0.3.0

# HuggingFace embeddings
sentence-transformers==3.0.1

# File parsing
pypdf==4.3.1
python-docx==1.1.2

# Gemini SDK (direct usage, latest)
google-generativeai==0.5.4

torch
transformers
